Split up codebase per executable
The source code for each of C based executables is now in a directory
named for the executable, with a common directory for the null-stream
code.

The make-package code has been further split up into different files
for function / group of functions, in order to reduce the size of each
source file for easier editing.