Move Installed-Size calculation outside of make-package
The information used in /var/lib/dpkg/status used the control file
included in the package.
Dpkg/apt use the Installed-Size as a kind of checksum to check that
the installed package is the same one listed in the package list,
even when the version numbers are the same. If this field is missing
from the in-package control file, the package shows as always
upgradable.

This is fixed by calculating the Installed-Size in the bpkg-build
wrapper script.
This is achieved backing up the control file, removing all the fields
that are set at any point by bpkg, and then writing back to the
main file with the Version and Installed-Size fields.
Once make-package exits (successfully or otherwise), the original
control file is put back in place.

The script is additionally updated to support adding a version at
this step, based on a second parameter. The Version must be specified
in either the CLI argument or in the control file.
If it is specified in both, the values must match for the package
to be built.