tar-stream: Add work arounds for different open(2) flags.
Different libc-s on different operating systems have different flag for the
`open(2)` family of functions.

In this code base, we use `O_PATH` to mean "only open the path/node, not the
file contents", and `O_SYMLINK` as "operate directly on a symlink".

On Linux `O_PATH` fills both of these purposes.
Some systems do not have the path-only flag.

For now, we will define `O_PATH` to `0` if not defined, then `O_SYMLINK`
to `O_PATH` (if not defined).